refactor(print-agent): Agent abgeben, Verträge behalten
tools/print_agent.py geht in Rente. Der Agent lebt neu in kp-rueck und spricht
dort beide Protokolle, damit eine Station, die beide Systeme betreibt, EINEN
Dienst betreibt statt zweier Agenten auf derselben Kiste für denselben
Druckerraum.

Was hier bleibt, ist das, was hierher gehört: der Endpunkt-Vertrag. Der Agent
wurde auf ihn portiert, nicht umgekehrt – Long-Poll-Claim, PDF-Download,
Statusmeldung, X-Print-Agent-Secret, alles unverändert. Der Ersatz ist deshalb
kein Link, sondern eine Seite, die den Vertrag nennt, die Migration beschreibt
und das portierte Verhalten festhält (wartender CUPS-Auftrag = ausstehend,
lp-Optionen nach den Defaults). Wer einen eigenen Agenten schreibt, hat weiterhin
alles hier.

Die Warnung steht bewusst als Warnung da: erst den alten Agenten stoppen. Zwei
Agenten an einer Warteschlange beanspruchen beide Aufträge, jeder Auftrag druckt
dann einmal – bei dem, der zuerst gefragt hat. Das Symptom sind Drucke, die
«manchmal nicht ankommen», während beide Logs gesund aussehen.

Dazu die zwei restlichen Angleichungen zwischen den Repos:

* RUNNING-BOTH.md liegt jetzt byte-identisch in beiden – eine Station stösst auf
  diese Probleme, bevor sie weiss, in welchem Repo sie nachschauen müsste. Neu
  darin die Zuordnungstabelle für Variablen, die dasselbe meinen und anders
  heissen (SECRET_KEY/AUTH_SECRET_KEY, COOKIE_SECURE/AUTH_COOKIE_SECURE,
  APP_PORT/HTTP_PORT, PRINT_AGENT_SECRET/PRINT_AGENT_TOKEN).
* RESERVED_ALARM_SOURCES ist neu die Vereinigung beider Listen. Wer eine
  Leitstelle in beide Apps führt, soll keinen Slug wählen können, den die eine
  annimmt und die andere mit 422 ablehnt – eine Falle, die erst bei der zweiten
  Integration zuschnappt. Einen Namen zu reservieren, den diese App nicht nutzt,
  kostet nichts: als externer Absendername war er ohnehin nie gültig.

+++ backend/app/weather.py
@@ -270,7 +270,7 @@ def _num(value: object) -> float | None:
 
 def _int(value: object) -> int | None:
     f = _num(value)
-    return None if f is None else int(round(f))
+    return None if f is None else round(f)
 
 
 weather_client = WeatherClient()
